+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [Unreleased]
 ### Added
-* Rgba32, Gray8 pixel formats.
+* Rgba8, Gray8 pixel formats.
 * Raster::width(), height() and pixels()
 ### Changed
 * Raster and Plotter now have pixel format type parameters.